+++ packages/generate-loader/index.js
@@ -1,10 +1,11 @@
 const yeoman = require("yeoman-environment");
-const { LoaderGenerator } = require("webpack-cli-generators/loader-generator");
+const { LoaderGenerator } = require("@webpack-cli/generators/loader-generator");
 
 /**
  * Runs a yeoman generator to create a new webpack loader project
  * @returns {void}
  */
+
 function loaderCreator() {
 	const env = yeoman.createEnv();
 	const generatorName = "webpack-loader-generator";