+++ packages/generators/plugin-generator.js
@@ -21,7 +21,7 @@ const PluginGenerator = webpackGenerator(
 			validate: str => str.length > 0
 		}
 	],
-	path.join(__dirname, "templates"),
+	path.resolve(__dirname, "..", "generate-plugin"),
 	[
 		"src/cjs.js.tpl",
 		"test/test-utils.js.tpl",
@@ -30,7 +30,10 @@ const PluginGenerator = webpackGenerator(
 		"examples/simple/src/lazy-module.js.tpl",
 		"examples/simple/src/static-esm-module.js.tpl"
 	],
-	["src/_index.js.tpl", "examples/simple/_webpack.config.js.tpl"],
+	[
+		"src/_index.js.tpl",
+		"examples/simple/_webpack.config.js.tpl"
+	],
 	gen => ({ name: _.upperFirst(_.camelCase(gen.props.name)) })
 );