+++ packages/migrate/moduleConcatenationPlugin/moduleConcatenationPlugin.test.js
@@ -1,0 +1,19 @@
+"use strict";
+
+const defineTest = require("@webpack-cli/utils/defineTest");
+
+defineTest(
+	__dirname,
+	"moduleConcatenationPlugin",
+	"moduleConcatenationPlugin-0"
+);
+defineTest(
+	__dirname,
+	"moduleConcatenationPlugin",
+	"moduleConcatenationPlugin-1"
+);
+defineTest(
+	__dirname,
+	"moduleConcatenationPlugin",
+	"moduleConcatenationPlugin-2"
+);