+++ packages/migrate/namedModulesPlugin/__testfixtures__/namedModulesPlugin-1.input.js
@@ -1,0 +1,9 @@
+module.export = {
+	optimizations: {
+		splitChunks: false
+	},
+    plugins: [
+        new Foo(),
+        new webpack.NamedModulesPlugin()
+    ]
+}