+++ packages/migrate/removeDeprecatedPlugins/removeDeprecatedPlugins.js
@@ -1,5 +1,5 @@
 const chalk = require("chalk");
-const utils = require("webpack-cli-utils/ast-utils");
+const utils = require("@webpack-cli/utils/ast-utils");
 
 /**
  *